Support packing directly into a writer object
When packing larger objects, the current default behavior
of packing the entire object into a buffer and then writing
it into the provided output stream results in massive slowdowns.

To address this case, introduce a StreamingPacker which takes
a callable, to which bytes are sent as they are produced. This
function will typically be the write() method of a file-like
object, but it can be any callable that accepts bytes.

The pure Python implementation writes directly to this callback.

The C implementation only writes to this function when its internal
buffer fills up, and to flush the remainder of the buffer at the
end of a pack*() call. Large elements may also bypass the buffer
completely.

Packer is redefined as a thin wrapper around StreamingPacker to
avoid code duplication while keeping StreamingPacker as simple
as possible.

Add a few more tests to cover large values