Support packing directly into a writer object
When packing larger objects, the current default behavior
of packing the entire object into a buffer and then writing
it into the provided output stream results in massive slowdowns.

To address this use case, Packer now accepts an optional `writer`
kwarg, which must be an object with a `write()` method. When this
argument is given:
  - the pure Python implementation writes directly to this object,
    completely bypassing StringIO
  - the C implementation only writes to this object when its internal
    buffer fills up, and to flush the remainder of the buffer at the
    end of a pack() call. Large elements may also bypass the buffer
    completely.

Add a few more tests to cover large values