Remove test_max_str_len
It appears this test was verifying a quirk of the DoS protection.
When the input is supplied via feed() BufferFull is raised to prevent
excessive memory allocation. When the input is supplied via a file
ValueError is raised when the length of the bin, str or ext is too big.

The fallback implementation had a quirk: it would raise ValueError when
the length of the is too big even when the input is supplied via feed().
As far as I can tell removing this quirk does not introduce a
vulnerability. I have fixed the quirk making this test obsolete.
This also allowed me to remove the code from the extension the appears
to serve only one purpose: make this test pass.