Do not fail silently when unpacking large messages
Fixes #503

Raising ValueError matches the behaviour of the pure Python version.
I'm not sure that this is the best place to raise this message; it might
make more sense when reading headers or in `append_buffer` or
something. It's also possible that we want to raise BufferFull instead
of ValueError.