docs: adjust comment about sys.path
note that it is usually not needed to tweak sys.path, one just needs to have
msgpack installed (like into a currently active venv) and it will be found.

i adjusted the path in the comment nevertheless, so if somebody wants the
stuff in docs/../msgpack to be found, it just needs to be uncommented.

sys.path tweaking might also cause issues i guess, e.g. if docs/../msgpack
is found first and stuff from there is tried, instead of a correctly installed
msgpack somewhere else in sys.path.